requirements.txt
boto3
botocore
flask
sentry_sdk
sentry-sdk[flask]
bullmq
requests
Pympler
python-dotenv==1.0.0
pyjwt
redis
bullmq